# docs/swagger.yaml
basePath: /
definitions:
  models.APIError:
    description: API error response
    properties:
      error:
        example: Internal server error
        type: string
    type: object
  models.CreateTodoInput:
    description: Input for creating a todo item
    properties:
      priority:
        enum:
        - 0
        - 1
        - 2
        - 3
        example: 1
        type: integer
      title:
        example: Buy milk
        type: string
    required:
    - title
    type: object
  models.Todo:
    description: Todo item model
    properties:
      completed:
        example: false
        type: boolean
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      priority:
        enum:
        - 0
        - 1
        - 2
        - 3
        example: 1
        type: integer
      title:
        example: Buy milk
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    required:
    - title
    type: object
  models.UpdateTodoInput:
    description: Input for updating a todo item
    properties:
      completed:
        example: true
        type: boolean
      priority:
        enum:
        - 0
        - 1
        - 2
        - 3
        example: 1
        type: integer
      title:
        example: Buy milk
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: A simple todo list API in Golang using Gin and GORM.
  title: Todo List API
  version: 1.0.0
paths:
  /api/v1/ping:
    get:
      consumes:
      - application/json
      description: v1 health check
      operationId: PingV1
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: V1 Health check
      tags:
      - health
  /api/v1/todos:
    get:
      consumes:
      - application/json
      description: get todos
      operationId: ListTodos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Todo'
            type: array
      summary: List todos
      tags:
      - todos
    post:
      consumes:
      - application/json
      description: create a new todo item
      operationId: CreateTodo
      parameters:
      - description: Todo object
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/models.CreateTodoInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Todo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Create a todo
      tags:
      - todos
  /api/v1/todos/{id}:
    delete:
      consumes:
      - application/json
      description: delete a todo item by ID
      operationId: DeleteTodo
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Delete a todo
      tags:
      - todos
    get:
      consumes:
      - application/json
      description: get a todo item by ID
      operationId: GetTodoByID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Todo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Get a todo
      tags:
      - todos
    put:
      consumes:
      - application/json
      description: update an existing todo item
      operationId: UpdateTodo
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated todo object
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTodoInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Todo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Update a todo
      tags:
      - todos
swagger: "2.0"